+++ packages/web/src/app/components/navigationMenu.tsx
@@ -7,8 +7,8 @@ import { GitHubLogoIcon, DiscordLogoIcon } from "@radix-ui/react-icons";
 import { SettingsDropdown } from "./settingsDropdown";
 import { Separator } from "@/components/ui/separator";
 import Image from "next/image";
-import logoDark from "../../public/sb_logo_dark_small.png";
-import logoLight from "../../public/sb_logo_light_small.png";
+import logoDark from "../../../public/sb_logo_dark_small.png";
+import logoLight from "../../../public/sb_logo_light_small.png";
 import { useRouter } from "next/navigation";
 
 const SOURCEBOT_DISCORD_URL = "https://discord.gg/6Fhp27x7Pb";
